Initial IR and Lua transform
Not happy with the IR. It should be a discriminated union of interfaces, but I wanted to try union of classes at least once since I did interfaces in my previous project.

Also we're going to need some sort of general traversal that also keeps track of parents for e.g. parentheses