Readonly IR (#48)
* Setup for search

* Super simple bad BFS search

* Add Spine and proper caching for BFS

* Support GolfPlugins in emitPlugins

* Ignore mutating visitors in golfing BFS

Removes the need to structuredClone everywhere

* Limit GolfPlugins to one replacement

* Migrate many plugins to GolfPlugin form

* Convert AST to readonly

Currently failing typecheck in 10 lines, and Path uses (as any) in what would otherwize fail typecheck

* Flip type inference to be query-based

* Cache types in WeakMap instead of on nodes

* Migrate more plugins to GolfPlugin form

* Migrate renameIdents to GolfPlugin

* Refactor weird sequential import insertion

Now there is no `.dependencies` on `Program`: the dependencies
are determined immediately before emit. This might be in need of
improvement for performance, similar to `symbolTableRoot`.

I removed the Python dependency thing because it doesn't seem to
be used at all. Will need to add back `sys` detection.

* Rewrite expandVariants to not use Path

* Remove Path

* Rename GolfPlugin to Plugin

`"lib": ["esnext"]` ensures that Plugin never accidentally resolves to the (deprecated) DOM `Plugin` type.

* Fix infinite recursion in getType via symbol table

* Rename getSymbolTable.ts to symbols.ts

* Fix typo in check for declaration assignments

* Add limit on BFS size

* Gracefully handle an error in the first emit

* Remove an unneeded clone

* Copy user-annotated types better

This only caused a bug in `$name: 0...5 <- 1` getting plugin transformed
to `$a <- 1` then all instances of `$a` getting constant-collapsed. But I
updated the others just to be safe.

Also you're lucky I didn't curry `getType`

* Pass node as first argument of Spine visitor

* Fix tests

* Update some comments

* Fix infinite recursion in type inference

* Expand comments and rename spine `visit` to `compactMap`

The rename is to clarify that the method is not an imperative-style
visitor. It's more like a functor map, with `undefined`s filtered out.

* Move around plugins for each language

The finalEmitPlugins can't get applied during golfing, just during emit

* Re-introduce check on variant count

* Copy type annotation from a node to its replacement

* Add back common/debug

Co-authored-by: MichalMarsalek <MichalMarsalek@users.noreply.github.com>