Handle for-range either in block or at root.
Reverts "revert forRangeToWhile block requirement"

Now, `forRangeToWhile` will only apply to a ForRange that is either the direct child of a block, or is the root node.

The test was failing because the plugin didn't apply to root node. But the old implementation using visit() directly didn't work because error "attempt to insert a block into a block".